NaArray-abind {SparseArray}	R Documentation

Combine multidimensional NaArray objects

Description

EXPERIMENTAL!!!

Like ordinary matrices and arrays in base R, NaMatrix objects can be combined by rows or columns, with rbind() or cbind(), and multidimensional NaArray objects can be bound along any dimension with abind().

Note that arbind() can also be used to combine the objects along their first dimension, and acbind() can be used to combine them along their second dimension.

See Also

cbind in base R.

abind in the S4Arrays package.

NaArray objects.

Ordinary array objects in base R.

Examples

# COMING SOON...

[Package SparseArray version 1.10.8 Index]